requirements.txt
djangorestframework
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
gunicorn==19.9.0
idna==2.8
isodate==0.6.0
psycopg2==2.8.3
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
